Revert previous fix for deprecated libcaca versions. It's possible we have a TOIlet release before a libcaca one.

+++ tools/caca2tlf.c
@@ -170,7 +170,7 @@ static void list_fonts(void)
 static void add_char(unsigned long int ch)
 {
     void *buf;
-    unsigned long int len;
+    size_t len;
     unsigned int x, y, myw, mygw;
     int full = caca_utf32_is_fullwidth(ch);
